+++ src/apps/settings/router.tsx
@@ -16,6 +16,8 @@ import { WindowManagerSettings } from './modules/WindowManager/main/infra';
 import { Layout } from './components/layout';
 import { RoutePath } from './components/navigation/routes';
 import { Home } from './modules/Home';
+import { ThemeView } from './modules/resources/Theme/View';
+import { WidgetView } from './modules/resources/Widget/View';
 
 export function Routing() {
   return (
@@ -34,7 +36,8 @@ export function Routing() {
         <Route path={RoutePath.WindowManager} Component={WindowManagerSettings} />
         <Route path={RoutePath.FancyToolbar} Component={FancyToolbarSettings} />
         <Route path={RoutePath.DevTools} Component={DeveloperTools} />
-        <Route path="widget/:username/:resourceName" Component={() => '🏗️🚧'} />
+        <Route path="widget/:username/:resourceName" Component={WidgetView} />
+        <Route path="theme/:username/:resourceName" Component={ThemeView} />
       </Route>
     </Routes>
   );