+++ changelog.md
@@ -5,6 +5,7 @@
 - add open window label on dock for app items (configurable).
 - add input to write custom text items on the toolbar.
 - add export resource button on developer tools.
+- allow settings by theme.
 
 ### enhancements
 - Settings UI refactor to follow the new resources ecosystem.